Moving license model to model.dart
