+++ lib/src/model.dart
@@ -10,7 +10,6 @@ import 'package:meta/meta.dart';
 import 'package:pub_semver/pub_semver.dart';
 
 import 'code_problem.dart';
-import 'license.dart';
 import 'platform.dart';
 import 'pubspec.dart';
 import 'sdk_info.dart';
@@ -617,4 +616,55 @@ class Fitness extends Object with _$FitnessSerializerMixin {
   }
 
   bool get hasSuggestion => suggestions != null && suggestions.isNotEmpty;
+}
+
+@JsonSerializable()
+class LicenseFile extends Object with _$LicenseFileSerializerMixin {
+  final String path;
+  final String name;
+
+  @JsonKey(includeIfNull: false)
+  final String version;
+
+  @JsonKey(includeIfNull: false)
+  final String url;
+
+  LicenseFile(this.path, this.name, {this.version, this.url});
+
+  factory LicenseFile.fromJson(Map<String, dynamic> json) =>
+      _$LicenseFileFromJson(json);
+
+  LicenseFile change({String url}) =>
+      new LicenseFile(path, name, version: version, url: url ?? this.url);
+
+  String get shortFormatted => version == null ? name : '$name $version';
+
+  @override
+  String toString() => '$path: $shortFormatted';
+
+  @override
+  bool operator ==(Object other) =>
+      identical(this, other) ||
+          other is LicenseFile &&
+              runtimeType == other.runtimeType &&
+              path == other.path &&
+              name == other.name &&
+              version == other.version &&
+              url == other.url;
+
+  @override
+  int get hashCode =>
+      path.hashCode ^ name.hashCode ^ version.hashCode ^ url.hashCode;
+}
+
+abstract class LicenseNames {
+  static const String AGPL = 'AGPL';
+  static const String Apache = 'Apache';
+  static const String BSD = 'BSD';
+  static const String GPL = 'GPL';
+  static const String LGPL = 'LGPL';
+  static const String MIT = 'MIT';
+  static const String MPL = 'MPL';
+  static const String Unlicense = 'Unlicense';
+  static const String unknown = 'unknown';
 }
